Fix icons #3 :: add white border + transparency
Instead of having a white background, icons now have a transparent
background outside the icon. A new solid element was added, behind the
black icon frame which serves as the **white background** and **border**
to the icon. This is similar to Google's PiP icon is.

Also, updated `Screenshot.png` to represent the new icon in the toolbar.